* migrate from byte to ByteBuffer for req/response * remove the tk package * move RequestLine to rq package
