+++ src/main/java/com/artipie/http/Connection.java
@@ -23,6 +23,7 @@
  */
 package com.artipie.http;
 
+import java.nio.ByteBuffer;
 import java.util.Map;
 import java.util.concurrent.Flow;
 
@@ -40,5 +41,5 @@ public interface Connection {
      */
     void accept(int code,
         Iterable<Map.Entry<String, String>> headers,
-        Flow.Publisher<Byte> body);
+        Flow.Publisher<ByteBuffer> body);
 }
