+++ src/main/java/com/artipie/http/Slice.java
@@ -23,6 +23,7 @@
  */
 package com.artipie.http;
 
+import java.nio.ByteBuffer;
 import java.util.Map;
 import java.util.concurrent.Flow;
 
@@ -47,5 +48,5 @@ public interface Slice {
      */
     Response response(String line,
         Iterable<Map.Entry<String, String>> headers,
-        Flow.Publisher<Byte> body);
+        Flow.Publisher<ByteBuffer> body);
 }
