+++ src/main/java/com/artipie/http/rq/RequestLine.java
@@ -21,9 +21,45 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+package com.artipie.http.rq;
 
 /**
- * Tests for Takes wrappers of Artipie.
+ * Http Request Line.
+ * <p>
+ * See: 5.1 https://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html
  * @since 0.1
  */
-package com.artipie.http.tk;
+public final class RequestLine {
+
+    /**
+     * The request method.
+     */
+    private final String method;
+
+    /**
+     * The request uri.
+     */
+    private final String uri;
+
+    /**
+     * The Http version.
+     */
+    private final String version;
+
+    /**
+     * Ctor.
+     * @param method The http method.
+     * @param uri The http uri.
+     * @param version The http version.
+     */
+    public RequestLine(final String method, final String uri, final String version) {
+        this.method = method;
+        this.uri = uri;
+        this.version = version;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("%s %s %s\r\n", this.method, this.uri, this.version);
+    }
+}
