+++ src/test/java/com/artipie/http/rq/RequestLineTest.java
@@ -21,45 +21,24 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package com.artipie.http;
+package com.artipie.http.rq;
+
+import org.hamcrest.MatcherAssert;
+import org.hamcrest.Matchers;
+import org.junit.jupiter.api.Test;
 
 /**
- * Http Request Line.
- * <p>
- * See: 5.1 https://www.w3.org/Protocols/rfc2616/rfc2616-sec5.html
+ * Ensure that {@link RequestLine} works correctly.
+ *
  * @since 0.1
  */
-public final class RequestLine {
+public class RequestLineTest {
 
-    /**
-     * The request method.
-     */
-    private final String method;
-
-    /**
-     * The request uri.
-     */
-    private final String uri;
-
-    /**
-     * The Http version.
-     */
-    private final String version;
-
-    /**
-     * Ctor.
-     * @param method The http method.
-     * @param uri The http uri.
-     * @param version The http version.
-     */
-    public RequestLine(final String method, final String uri, final String version) {
-        this.method = method;
-        this.uri = uri;
-        this.version = version;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("%s %s %s\r\n", this.method, this.uri, this.version);
+    @Test
+    public void reqLineStringIsCorrect() {
+        MatcherAssert.assertThat(
+            new RequestLine("GET", "/pub/WWW/TheProject.html", "HTTP/1.1").toString(),
+            Matchers.equalTo("GET /pub/WWW/TheProject.html HTTP/1.1\r\n")
+        );
     }
 }
